feat(static-default-options): 获取默认配置项设置为静态方法
- [x] line,column,bar,area,tiny-line/column/area,ring-progress,progress （主要供 multi-view 使用）
