+++ src/plots/area/index.ts
@@ -1,36 +1,29 @@
 import { Plot } from '../../core/plot';
-import { deepAssign } from '../../utils';
 import { Adaptor } from '../../core/adaptor';
 import { getDataWhetherPecentage } from '../../utils/transform/percent';
 import { AreaOptions } from './types';
 import { adaptor } from './adaptor';
+import { DEFAULT_OPTIONS } from './constants';
 
 export { AreaOptions };
 
 export class Area extends Plot<AreaOptions> {
+  /**
+   * 获取 面积图 默认配置项
+   * @static 供外部使用
+   */
+  static getDefaultOptions(): Partial<AreaOptions> {
+    return DEFAULT_OPTIONS;
+  }
+
   /** 图表类型 */
   public type: string = 'area';
 
   /**
-   * 获取 折线图 默认配置
+   * 获取 面积图 默认配置
    */
   protected getDefaultOptions() {
-    return deepAssign({}, super.getDefaultOptions(), {
-      tooltip: {
-        shared: true,
-        showMarkers: true,
-        showCrosshairs: true,
-        crosshairs: {
-          type: 'x',
-        },
-      },
-      isStack: true,
-      // 默认开启
-      line: {},
-      legend: {
-        position: 'top-left',
-      },
-    });
+    return Area.getDefaultOptions();
   }
 
   /**
