+++ src/plots/gauge/index.ts
@@ -2,8 +2,9 @@ import { Plot } from '../../core/plot';
 import { Adaptor } from '../../core/adaptor';
 import { GaugeOptions } from './types';
 import { adaptor, statistic } from './adaptor';
-import { RANGE_VALUE, PERCENT, INDICATEOR_VIEW_ID, RANGE_VIEW_ID } from './constant';
+import { INDICATEOR_VIEW_ID, RANGE_VIEW_ID, DEFAULT_OPTIONS } from './constants';
 import { getIndicatorData, getRangeData } from './utils';
+
 // 注册 shape
 import './shapes/gauge';
 import './shapes/meter-gauge';
@@ -11,72 +12,19 @@ import './shapes/meter-gauge';
 export { GaugeOptions };
 
 /**
- * 仪表盘盘
+ * 仪表盘
  */
 export class Gauge extends Plot<GaugeOptions> {
+  /**
+   * 获取 仪表盘 默认配置项
+   * @static 供外部使用
+   */
+  static getDefaultOptions(): Partial<GaugeOptions> {
+    return DEFAULT_OPTIONS;
+  }
+
   /** 图表类型 */
   public type: string = 'gauge';
-
-  protected getDefaultOptions() {
-    return {
-      percent: 0, // 当前指标值
-      range: {
-        ticks: [],
-      }, // 默认的刻度
-      innerRadius: 0.9,
-      radius: 0.95,
-      startAngle: (-7 / 6) * Math.PI,
-      endAngle: (1 / 6) * Math.PI,
-      syncViewPadding: true,
-      axis: {
-        line: null,
-        label: {
-          offset: -24,
-          style: {
-            textAlign: 'center',
-            textBaseline: 'middle',
-          },
-        },
-        subTickLine: {
-          length: -8,
-        },
-        tickLine: {
-          length: -12,
-        },
-        grid: null,
-      },
-      indicator: {
-        pointer: {
-          style: {
-            lineWidth: 5,
-            lineCap: 'round',
-          },
-        },
-        pin: {
-          style: {
-            r: 9.75,
-            lineWidth: 4.5,
-            fill: '#fff',
-          },
-        },
-      },
-      statistic: {
-        title: false,
-      },
-      meta: {
-        // 两个 view 的 scale 同步到 v 上
-        [RANGE_VALUE]: {
-          sync: 'v',
-        },
-        [PERCENT]: {
-          sync: 'v',
-          tickCount: 5,
-          tickInterval: 0.2,
-        },
-      },
-      animation: false,
-    };
-  }
 
   /**
    * 更新数据
@@ -96,6 +44,14 @@ export class Gauge extends Plot<GaugeOptions> {
     }
     // todo 后续让 G2 层在 afterrender 之后，来重绘 annotations
     statistic({ chart: this.chart, options: this.options }, true);
+  }
+
+  /**
+   * 获取默认配置
+   * 供 base 使用
+   */
+  protected getDefaultOptions() {
+    return Gauge.getDefaultOptions();
   }
 
   /**
