--- src/plots/progress/constant.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_COLOR = ['#FAAD14', '#E8EDF3'];
