+++ src/plots/ring-progress/index.ts
@@ -3,35 +3,21 @@ import { Adaptor } from '../../core/adaptor';
 import { getProgressData } from '../progress/utils';
 import { RingProgressOptions } from './types';
 import { adaptor, statistic } from './adaptor';
+import { DEFAULT_OPTIONS } from './constants';
 
 export { RingProgressOptions };
 
 export class RingProgress extends Plot<RingProgressOptions> {
+  /**
+   * 获取默认配置项
+   * @static 供外部使用
+   */
+  static getDefaultOptions(): Partial<RingProgressOptions> {
+    return DEFAULT_OPTIONS;
+  }
+
   /** 图表类型 */
   public type: string = 'ring-process';
-
-  protected getDefaultOptions() {
-    return {
-      percent: 0.2,
-      innerRadius: 0.8,
-      radius: 0.98,
-      color: ['#FAAD14', '#E8EDF3'],
-      statistic: {
-        title: false,
-        content: {
-          style: {
-            fontSize: '14px',
-            fontWeight: 300,
-            fill: '#4D4D4D',
-            textAlign: 'center',
-            textBaseline: 'middle',
-          },
-          formatter: ({ percent }) => `${(percent * 100).toFixed(2)}%`,
-        },
-      },
-      animation: true,
-    };
-  }
 
   /**
    * 更新数据
@@ -43,6 +29,10 @@ export class RingProgress extends Plot<RingProgressOptions> {
     this.chart.data(getProgressData(percent));
     // todo 后续让 G2 层在 afterrender 之后，来重绘 annotations
     statistic({ chart: this.chart, options: this.options }, true);
+  }
+
+  protected getDefaultOptions() {
+    return RingProgress.getDefaultOptions();
   }
 
   /**
