+++ src/plots/tiny-column/index.ts
@@ -2,12 +2,20 @@ import { Adaptor } from '../../core/adaptor';
 import { Plot } from '../../core/plot';
 import { getTinyData } from '../tiny-line/utils';
 import { adaptor } from './adaptor';
-import { DEFAULT_TOOLTIP_OPTIONS } from './constants';
+import { DEFAULT_OPTIONS } from './constants';
 import { TinyColumnOptions } from './types';
 
 export { TinyColumnOptions };
 
 export class TinyColumn extends Plot<TinyColumnOptions> {
+  /**
+   * 获取默认配置项
+   * @static 供外部使用
+   */
+  static getDefaultOptions(): Partial<TinyColumnOptions> {
+    return DEFAULT_OPTIONS;
+  }
+
   /** 图表类型 */
   public type: string = 'tiny-column';
 
@@ -21,13 +29,7 @@ export class TinyColumn extends Plot<TinyColumnOptions> {
   }
 
   protected getDefaultOptions() {
-    return {
-      appendPadding: 2,
-      tooltip: {
-        ...DEFAULT_TOOLTIP_OPTIONS,
-      },
-      animation: true,
-    };
+    return TinyColumn.getDefaultOptions();
   }
 
   /**
