+++ src/plots/tiny-line/constants.ts
@@ -18,6 +18,17 @@ export const DEFAULT_TOOLTIP_OPTIONS = {
   },
   showCrosshairs: true,
   crosshairs: {
-    type: 'x',
+    type: 'x' as const,
   },
+};
+
+/**
+ * 默认配置项
+ */
+export const DEFAULT_OPTIONS = {
+  appendPadding: 2,
+  tooltip: {
+    ...DEFAULT_TOOLTIP_OPTIONS,
+  },
+  animation: {},
 };
