+++ src/types/statistic.ts
@@ -1,8 +1,9 @@
 import { View } from '@antv/g2';
 import { Data, Datum } from './common';
 
-type CSSStyle = Omit<Partial<CSSStyleDeclaration>, 'opacity' | 'lineHeight'> & {
+type CSSStyle = Omit<Partial<CSSStyleDeclaration>, 'opacity' | 'fontWeight' | 'lineHeight'> & {
   opacity?: number;
+  fontWeight?: string | number;
   lineHeight?: string | number;
 };
